fix: Namespace ModeType and fix tests
