+++ app/graphql/types/query_type.rb
@@ -6,7 +6,7 @@ module Types
     # Add root-level fields here.
     # They will be entry points for queries on your schema.
 
-    field :journey_meta_modes, [ModeType], null: false,
+    field :journey_meta_modes, [Tfl::Entities::ModeType], null: false,
                                description: 'Gets a list of all of the available journey planner modes'
     def journey_meta_modes
       @client = TflApi::Client.new(app_id: ENV['TFL_APP_ID'], app_key: ENV['TFL_APP_KEY'], host: ENV['TFL_APP_BASE'])
